requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
docarray==0.41.0
docker==7.1.0
ecdsa==0.19.1
einops==0.8.1
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
grpcio==1.68.0
grpcio-health-checking==1.68.0
grpcio-reflection==1.68.0
h11==0.14.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
jcloud==0.3
jina==3.34.0
jina-hubble-sdk==0.39.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
narwhals==1.36.0
networkx==3.2.1
numpy==2.0.2
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-exporter-prometheus==0.53b1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-aiohttp-client==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-instrumentation-grpc==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
prometheus_client==0.21.1
propcache==0.3.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.4.8
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
streamlit==1.44.1
sympy==1.13.1
tenacity==9.1.2
timm==1.0.15
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.23.1
uvloop==0.21.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
